requirements.txt
python=3.12
pandas
matplotlib
tqdm
pytorch
seaborn
scikit-learn
requests